[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17688 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,11 +4,10 @@ public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		
-		int a = sc.nextInt();
-		int b = sc.nextInt();
+		long a = sc.nextInt();
+		long b = sc.nextInt();
 		
 		long result = (a+b) * (a-b);
 		System.out.println(result);
 	}
-
 }
